requirements.txt
ipykernel
python-dotenv
pandas
numpy
yfinance
scikit-learn
ta
matplotlib
joblib
tensorflow
pyarrow